+++ packages/ts-utils/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 0.7.0
 
-Initial creation and publication of this package, as an extraction and de-duplication from `@lionweb/core`, `@lionweb/utilities`, `@lionweb/class-core`, and `@lionweb/validation`.
+* Initial creation and publication of this package, as an extraction and de-duplication from `@lionweb/core`, `@lionweb/utilities`, `@lionweb/class-core`, and `@lionweb/validation`.
+* Add 'mapFrom' function that maps an array to a map, using given key and value functions.